+++ device_preview/lib/device_preview.dart
@@ -1,3 +1,4 @@
 export 'src/devices/devices.dart';
 export 'src/device_preview.dart';
+export 'src/device_preview_data.dart';
 export 'src/screenshots/screenshot.dart';